+++ packages/client-rpc/src/rpc_interfaces.ts
@@ -48,7 +48,7 @@ export interface ParcnetGPCRPC {
 export interface ParcnetIdentityRPC {
   getSemaphoreV3Commitment: () => Promise<bigint>;
   getSemaphoreV4Commitment: () => Promise<bigint>;
-  getSemaphoreV4PublicKey: () => Promise<string>;
+  getPublicKey: () => Promise<string>;
 }
 
 export interface ParcnetPODRPC {