+++ examples/test-app/CHANGELOG.md
@@ -1,5 +1,13 @@
 # test-app
 
+## 1.0.3
+
+### Patch Changes
+
+- Rename getSemaphoreV4PublicKey -> getPublicKey
+- Updated dependencies
+  - @parcnet-js/app-connector@0.0.3
+
 ## 1.0.2
 
 ### Patch Changes